Save Commit

+++ ft_printf/rendu_ft_printf/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: mgirardo <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/17 18:10:12 by mgirardo          #+#    #+#             */
-/*   Updated: 2022/01/17 18:11:32 by mgirardo         ###   ########.fr       */
+/*   Updated: 2022/01/18 16:31:12 by mgirardo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftptf.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
 	size_t	i;
